requirements.txt
Flask
mysql-connector-python
faiss-cpu
sentence-transformers
transformers
numpy